fix: rename entry point
